Add AbstractIntegrationTest for TestContainers usage
This commit adds a new class called AbstractIntegrationTest, which any test class that extends it will automatically use TestContainers. If a developer wants to use a local Postgres database instead, they simply don't have to extend this class.

refactor: Remove containers package from test package

Previously, we had a containers folder in our test package, but since we will only be using Postgres in this application, we don't need this folder for any future use. This commit removes the containers folder in accordance with the YAGNI (You Aren't Gonna Need It) principle.